requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
jinja2==3.1.6
python-multipart==0.0.20
aiogram==3.7.0
python-dotenv==1.0.1
itsdangerous==2.2.0
